fixed updateChatsBasedOnConnectedDevices

+++ app/src/main/java/com/flydrop2p/flydrop2p/data/local/ChatInfoDAO.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface ChatInfoDAO {
     @Query("SELECT * FROM ChatInfoEntity WHERE id = :chatId")
-    fun getChatInfoById(chatId: Int): Flow<ChatInfoEntity>
+    fun getChatInfoById(chatId: Long): Flow<ChatInfoEntity>
 
     @Query("SELECT * FROM ChatInfoEntity")
     fun getChatsInfo(): Flow<List<ChatInfoEntity>>
